pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "discord-py[voice]>=2.4.0",
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "openai>=1.61.1",
    "python-dotenv>=1.0.1",
    "sendgrid>=6.11.0",
    "twilio>=9.4.4",
    "supervisor>=4.2.5",
    "pynacl>=1.5.0",
    "yt-dlp>=2025.1.26",
    "spotipy>=2.25.0",
    "aiohttp>=3.11.12",
    "slack-sdk>=3.34.0",
    "lyricsgenius>=3.2.0",
    "trafilatura>=2.0.0",
    "requests>=2.32.3",
]
